docs: Rename project from susstral to CraftyLlamaBot
The project has been renamed from 'susstral' to 'CraftyLlamaBot'. All instances of the old name have been replaced with the new name in the project files. Additionally, the project description in the pyproject.toml file has been updated to reflect the new purpose of the project as a Discord bot for the CraftyLlamaCoalition.